@@ -4,7 +4,7 @@ requires = [
     "pybind11==2.13.6",
     "pybind11_stubgen @ git+https://github.com/sarlinpe/pybind11-stubgen@sarlinpe/fix-2024-11",
     "numpy",
-    "ruff==0.6.7",
+    "ruff==0.12.7",
     "clang-format==19.1.0",
 ]
 build-backend = "scikit_build_core.build"
@@ -26,7 +26,7 @@ authors = [
 ]
 license = {text = "BSD-3-Clause"}
 urls = {Repository = "https://github.com/colmap/colmap"}
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = ["numpy"]
 classifiers = [
   "License :: OSI Approved :: BSD License",
@@ -39,10 +39,10 @@ cmake.source-dir = "python/"
 wheel.packages = ["python/pycolmap"]
 
 [tool.cibuildwheel]
-build = "cp3{8,9,10,11,12,13}-{macosx,manylinux,win}*"
+build = "cp3{9,10,11,12,13,14}-{macosx,manylinux,win}*"
 archs = ["auto64"]
-test-requires = "mypy==1.12.1"
-test-command = "python -c \"import pycolmap; print(pycolmap.__version__)\" &&  python -m mypy --package pycolmap --implicit-optional"
+test-requires = "pytest mypy==1.17.0 enlighten==1.13.0"
+test-command = "python -c \"import pycolmap; print(pycolmap.__version__)\" &&  python -m mypy --package pycolmap --implicit-optional && pytest {project}/python/examples/custom_incremental_pipeline_test.py"
 
 [tool.cibuildwheel.environment]
 VCPKG_COMMIT_ID = "0cb95c860ea83aafc1b24350510b30dec535989a"
