@@ -44,7 +44,7 @@ def parse_args():
     parser.add_argument(
         "--app_path",
         required=True,
-        help="The application path, e.g., " "build/COLMAP-dev-windows",
+        help="The application path, e.g., build/COLMAP-dev-windows",
     )
     args = parser.parse_args()
     return args
