@@ -102,9 +102,7 @@ def make2d(array, cols=None, dtype=None):
 
     """
     if (cols is None or dtype is None) and not len(array):
-        raise RuntimeError(
-            "cols and dtype must be specified for empty " "array"
-        )
+        raise RuntimeError("cols and dtype must be specified for empty array")
 
     if cols is None:
         cols = len(array[0])
@@ -532,7 +530,7 @@ class PlyElement(object):
             raise TypeError("only numpy arrays are supported")
 
         if len(data.shape) != 1:
-            raise ValueError("only one-dimensional arrays are " "supported")
+            raise ValueError("only one-dimensional arrays are supported")
 
         count = len(data)
 
@@ -553,7 +551,7 @@ class PlyElement(object):
                 if t[1][1] == "O":
                     if len(t) != 2:
                         raise ValueError(
-                            "non-scalar object fields not " "supported"
+                            "non-scalar object fields not supported"
                         )
 
                 len_str = _data_type_reverse[len_types.get(t[0], "u1")]
@@ -761,17 +759,17 @@ class PlyProperty(object):
 
         if line[1] == "list":
             if len(line) > 5:
-                raise PlyParseError("too many fields after " "'property list'")
+                raise PlyParseError("too many fields after 'property list'")
             if len(line) < 5:
-                raise PlyParseError("too few fields after " "'property list'")
+                raise PlyParseError("too few fields after 'property list'")
 
             return PlyListProperty(line[4], line[2], line[3])
 
         else:
             if len(line) > 3:
-                raise PlyParseError("too many fields after " "'property'")
+                raise PlyParseError("too many fields after 'property'")
             if len(line) < 3:
-                raise PlyParseError("too few fields after " "'property'")
+                raise PlyParseError("too few fields after 'property'")
 
             return PlyProperty(line[2], line[1])
 
