@@ -97,9 +97,12 @@ void SynthesizeExhaustiveMatches(double inlier_match_ratio,
       AddOutlierMatches(
           inlier_match_ratio, num_points2D1, num_points2D2, &matches);
 
-      database->WriteMatches(image1.ImageId(), image2.ImageId(), matches);
-      database->WriteTwoViewGeometry(
-          image1.ImageId(), image2.ImageId(), two_view_geometry);
+      if (!database->ExistsMatches(image_id1, image_id2)) {
+        database->WriteMatches(image_id1, image_id2, matches);
+      }
+      if (!database->ExistsInlierMatches(image_id1, image_id2)) {
+        database->WriteTwoViewGeometry(image_id1, image_id2, two_view_geometry);
+      }
     }
   }
 }
@@ -134,32 +137,34 @@ void SynthesizeChainedMatches(double inlier_match_ratio,
     }
   }
 
-  for (auto& two_view_geometry : two_view_geometries) {
-    const auto image_pair =
-        Database::PairIdToImagePair(two_view_geometry.first);
-    const auto& image1 = reconstruction->Image(image_pair.first);
+  for (auto& [pair_id, two_view_geometry] : two_view_geometries) {
+    const auto [image_id1, image_id2] = Database::PairIdToImagePair(pair_id);
+    const auto& image1 = reconstruction->Image(image_id1);
     const auto& camera1 = *image1.CameraPtr();
-    const auto& image2 = reconstruction->Image(image_pair.second);
+    const auto& image2 = reconstruction->Image(image_id2);
     const auto& camera2 = *image2.CameraPtr();
-    two_view_geometry.second.config = TwoViewGeometry::CALIBRATED;
-    two_view_geometry.second.cam2_from_cam1 =
+    two_view_geometry.config = TwoViewGeometry::CALIBRATED;
+    two_view_geometry.cam2_from_cam1 =
         image2.CamFromWorld() * Inverse(image1.CamFromWorld());
-    two_view_geometry.second.E =
-        EssentialMatrixFromPose(two_view_geometry.second.cam2_from_cam1);
-    two_view_geometry.second.F =
+    two_view_geometry.E =
+        EssentialMatrixFromPose(two_view_geometry.cam2_from_cam1);
+    two_view_geometry.F =
         FundamentalFromEssentialMatrix(camera2.CalibrationMatrix(),
-                                       two_view_geometry.second.E,
+                                       two_view_geometry.E,
                                        camera1.CalibrationMatrix());
 
-    FeatureMatches matches = two_view_geometry.second.inlier_matches;
+    FeatureMatches matches = two_view_geometry.inlier_matches;
     AddOutlierMatches(inlier_match_ratio,
                       image1.NumPoints2D(),
                       image2.NumPoints2D(),
                       &matches);
 
-    database->WriteMatches(image1.ImageId(), image2.ImageId(), matches);
-    database->WriteTwoViewGeometry(
-        image1.ImageId(), image2.ImageId(), two_view_geometry.second);
+    if (!database->ExistsMatches(image_id1, image_id2)) {
+      database->WriteMatches(image_id1, image_id2, matches);
+    }
+    if (!database->ExistsInlierMatches(image_id1, image_id2)) {
+      database->WriteTwoViewGeometry(image_id1, image_id2, two_view_geometry);
+    }
   }
 }
 
@@ -183,9 +188,12 @@ void SynthesizeDataset(const SyntheticDatasetOptions& options,
   }
 
   // Synthesize 3D points on unit sphere centered at origin.
+  std::unordered_set<point3D_t> new_points3D_ids;
+  new_points3D_ids.reserve(options.num_points3D);
   for (int point3D_idx = 0; point3D_idx < options.num_points3D; ++point3D_idx) {
-    reconstruction->AddPoint3D(Eigen::Vector3d::Random().normalized(),
-                               /*track=*/{});
+    new_points3D_ids.insert(
+        reconstruction->AddPoint3D(Eigen::Vector3d::Random().normalized(),
+                                   /*track=*/{}));
   }
 
   int total_num_images = (database == nullptr) ? 0 : database->NumImages();
@@ -300,8 +308,14 @@ void SynthesizeDataset(const SyntheticDatasetOptions& options,
         points2D.reserve(options.num_points3D +
                          options.num_points2D_without_point3D);
 
-        // Create 3D point observations by project all 3D points to the image.
+        // Create 3D point observations by projecting 3D points to the image.
         for (auto& [point3D_id, point3D] : reconstruction->Points3D()) {
+          if (new_points3D_ids.count(point3D_id) == 0) {
+            // If a non-empty reconstruction is given, only add tracks for newly
+            // added images and 3D points.
+            continue;
+          }
+
           Point2D point2D;
           const std::optional<Eigen::Vector2d> proj_point2D =
               camera.ImgFromCam(cam_from_world * point3D.xyz);
